Vagrant-Box/requirements.txt
pydantic==1.9.0
tensorfn==0.1.23
torch>=2.0.0+cpu
torchvision
opencv-python
lmdb
Shapely
Pillow
image
black
numpy
rapidfuzz
Polygon3
editdistance
asyncvnc